@@ -1,7 +1,7 @@
 package main
 
 // An example illustrating how to pipe in data to a Bubble Tea application.
-// Moreso, this serves as proof that Bubble Tea will automatically listen for
+// More so, this serves as proof that Bubble Tea will automatically listen for
 // keystrokes when input is not a TTY, such as when data is piped or redirected
 // in.
 
